Support analog and digital messages. Further define the Firmata command byte logic to handle 'channels' or pins/ports which get padded along with the command

+++ src/main/java/com/bortbort/arduino/FiloFirmata/Parser/Builders/SysexReportFirmwareBuilder.java
@@ -59,8 +59,8 @@ public class SysexReportFirmwareBuilder extends SysexMessageBuilder {
         }
 
         return new SysexReportFirmwareMessage(
-                messageBody[0],
-                messageBody[1],
+                (int) messageBody[0],
+                (int) messageBody[1],
                 firmwareName
         );
     }
